requirements.txt
pytest
selenium
webdriver-manager
allure-pytest
selenium-stealth
fake-useragent
